refactoring auth into userdata store
