+++ src/store/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import { auth } from './auth';
+import { userData } from './userData';
 import { login } from './login';
 
 // Here, we're combining all reducers we'll make into a single reducer that is being exported to index.js
@@ -7,6 +7,6 @@ import { login } from './login';
 // we can break out each reducer into it's own separate file and place between the curlies below.
 
 export const rootReducer = combineReducers({
-  auth,
+  userData,
   login
 });
